requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.9.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
frozenlist==1.4.1
gql==3.5.0
graphql-core==3.2.4
h11==0.14.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
propcache==0.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.20
wsproto==1.2.0
yarl==1.15.2
